Refactor examples folder organization and cleanup
- Organized examples into categorical subdirectories: basic, quadrotor, i_mppi, autotuning, cuda, and benchmarks.
- Moved and renamed 15 example files to their respective categories.
- Removed redundant root-level i_mppi_simulation.py and legacy i_mppi_simulation_legacy.py.
- Updated sys.path adjustments in example files to correctly locate the project root.
- Updated import statements in tests and example files to match the new structure.
- Updated justfile with new example paths.
- Comprehensive documentation update with a new root examples/README.md and category-specific READMEs.
