+++ examples/autotuning/README.md
@@ -1,0 +1,7 @@
+# Autotuning Examples
+
+Examples showing how to automatically optimize MPPI hyperparameters.
+
+- `basic.py`: Minimal setup for tuning `lambda` and `noise_sigma`.
+- `pendulum.py`: Tuning a controller for robust pendulum swing-up.
+- `evosax_comparison.py`: Benchmarking different JAX-native evolutionary strategies for parameter optimization.
