+++ examples/basic/README.md
@@ -1,0 +1,5 @@
+# Basic Examples
+
+Simple introductory examples for JAX-MPPI.
+
+- `pendulum.py`: Classic inverted pendulum swing-up and stabilization task. Demonstrates basic MPPI configuration, dynamics definition, and cost function implementation.
