+++ examples/benchmarks/README.md
@@ -1,0 +1,5 @@
+# Benchmarks
+
+Comparative analysis and performance testing.
+
+- `smooth_comparison.py`: Evaluates the smoothness and control effort of different MPPI variants (MPPI vs SMPPI vs KMPPI) on a navigation task.
