+++ examples/benchmarks/smooth_comparison.py
