+++ examples/cuda/README.md
@@ -1,0 +1,8 @@
+# CUDA Acceleration Examples
+
+Examples and tests for the CUDA-accelerated MPPI implementation.
+
+- `pendulum_jit.py`: Shows how to define and JIT-compile custom dynamics for the CUDA backend.
+- `test_cuda_config.py`: Verifies CUDA configuration.
+- `test_cuda_bindings.py`: Tests the Python-C++ bindings.
+- `test_cuda_mppi.py`: Integrated test for CUDA-based MPPI.
