+++ examples/i_mppi/README.md
@@ -1,0 +1,6 @@
+# Informative MPPI (I-MPPI) Examples
+
+Examples of informative control for exploration and mapping.
+
+- `simulation.py`: Full simulation of the two-layer I-MPPI architecture (Zhang et al. 2020), combining global path planning with local informative reactivity.
+- `fsmi_grid_demo.py`: Visualizes the Fast Shannon Mutual Information (FSMI) calculation on occupancy grids.
