+++ examples/quadrotor/README.md
@@ -1,0 +1,10 @@
+# Quadrotor Examples
+
+Trajectory tracking and control for quadrotors using JAX-MPPI.
+
+- `hover.py`: Stabilizes a quadrotor at a fixed altitude.
+- `circle.py`: Follows a circular trajectory in the horizontal plane.
+- `figure8.py`: Performs an aggressive figure-8 maneuver and compares MPPI, SMPPI, and KMPPI.
+- `custom_trajectory.py`: Demonstrates following smooth trajectories generated from user-defined waypoints.
+- `hover_comparison.py`: Comparative analysis of MPPI variants on a simple hover task.
+- `trajectories.py`: Utility functions for generating various reference trajectories.
