+++ examples/quadrotor/__init__.py
@@ -1,6 +1,6 @@
 """Quadrotor trajectory following examples."""
 
-from examples.quadrotor.trajectories import (
+from .trajectories import (
     generate_circle_trajectory,
     generate_hover_setpoint,
     generate_lemniscate_trajectory,
