+++ examples/quadrotor/figure8.py
@@ -14,12 +14,12 @@ import time
 from pathlib import Path
 
 # Add parent directory to path for imports when running directly
-sys.path.insert(0, str(Path(__file__).parent.parent))
+sys.path.insert(0, str(Path(__file__).parent.parent.parent))
 
 import jax
 import jax.numpy as jnp
 
-from examples.quadrotor.trajectories import (
+from .trajectories import (
     compute_trajectory_metrics,
     generate_lemniscate_trajectory,
 )
@@ -613,7 +613,7 @@ def run_quadrotor_figure8_comparison(
             plt.tight_layout()
 
             # Save
-            output_dir = Path(__file__).parent.parent / "docs" / "media"
+            output_dir = Path(__file__).parent.parent.parent / "docs" / "media"
             output_dir.mkdir(parents=True, exist_ok=True)
             output_path = output_dir / "quadrotor_figure8_comparison.png"
